* Renamed a file. * Moved amplitudes in RoughAmplitude to a run-time variable.
